feat: test negative case with wrong ids.

+++ src/usecases/delete_task.py
@@ -2,4 +2,7 @@ from persistence.csv_repository import CsvRepository
 
 
 def delete_task(id: str) -> None:
-    CsvRepository.get_instance().delete_task(id)
+    if CsvRepository.get_instance().task_exists(id):
+        CsvRepository.get_instance().delete_task(id)
+    else:
+        raise Exception("Task does not exist!")
